drop the old openaiauth
